+++ src/QuickMarcEditor/QuickMarcEditorRows/QuickMarcEditorRows.js
@@ -15,7 +15,10 @@ import {
 import isEqual from 'lodash/isEqual';
 import defer from 'lodash/defer';
 
-import { Pluggable } from '@folio/stripes/core';
+import {
+  Pluggable,
+  IfPermission,
+} from '@folio/stripes/core';
 import {
   TextField,
   Tooltip,
@@ -29,6 +32,7 @@ import { MaterialCharsField } from './MaterialCharsField';
 import { PhysDescriptionField } from './PhysDescriptionField';
 import { FixedFieldFactory } from './FixedField';
 import { LocationField } from './LocationField';
+import { DeletedRowPlaceholder } from './DeletedRowPlaceholder';
 import {
   isReadOnly,
   hasIndicatorException,
@@ -57,6 +61,7 @@ const QuickMarcEditorRows = ({
     markRecordDeleted,
     deleteRecord,
     moveRecord,
+    restoreRecord,
   },
   marcType,
 }) => {
@@ -105,7 +110,7 @@ const QuickMarcEditorRows = ({
 
       prevDeleteIcon?.focus();
     });
-  }, [fields, deleteRecord, markRecordDeleted, isNewRow]);
+  }, [fields, deleteRecord, markRecordDeleted, isNewRow, containerRef]);
 
   const moveRow = useCallback(({ target }) => {
     moveRecord({
@@ -113,6 +118,10 @@ const QuickMarcEditorRows = ({
       indexToSwitch: parseInt(target.dataset.indexToSwitch, 10),
     });
   }, [moveRecord]);
+
+  const restoreRow = useCallback((index) => {
+    restoreRecord({ index });
+  }, [restoreRecord]);
 
   const processTagRef = useCallback(ref => {
     if (!ref) return;
@@ -142,7 +151,12 @@ const QuickMarcEditorRows = ({
             }
 
             if (recordRow._isDeleted) {
-              return null;
+              return (
+                <DeletedRowPlaceholder
+                  field={recordRow}
+                  restoreRow={() => restoreRow(idx)}
+                />
+              );
             }
 
             const isDisabled = isReadOnly(recordRow, action, marcType);
@@ -370,11 +384,11 @@ const QuickMarcEditorRows = ({
                   )
                 }
 
-                <Pluggable
-                  type="find-authority"
-                >
-                  <FormattedMessage id="ui-quick-marc.noPlugin" />
-                </Pluggable>
+                <IfPermission perm="ui-quick-marc.quick-marc-authority-records.linkUnlink">
+                  <Pluggable type="find-authority">
+                    <FormattedMessage id="ui-quick-marc.noPlugin" />
+                  </Pluggable>
+                </IfPermission>
               </div>
             );
           })
@@ -401,6 +415,7 @@ QuickMarcEditorRows.propTypes = {
     deleteRecord: PropTypes.func.isRequired,
     markRecordDeleted: PropTypes.func.isRequired,
     moveRecord: PropTypes.func.isRequired,
+    restoreRecord: PropTypes.func.isRequired,
   }),
   marcType: PropTypes.oneOf(Object.values(MARC_TYPES)).isRequired,
 };